Start the shared `argparse_` module
This `argparse_` module is expected to house custom helpers for
interacting/extending the `argparse` standard library.

Move `add_default_command()` to a shared `argparse_` module so that
it can be used by commands without having circular import issues.

I named the module `argparse_` (with an underscore) to avoid overwriting
the `argparse` standard library within `augur/__init__.py`. @tsibley
found the relevant docs¹ to explain this behavior.

¹ https://docs.python.org/3/reference/import.html#submodules
