+++ augur/argparse_.py
@@ -1,0 +1,16 @@
+"""
+Custom helpers for the argparse standard library.
+"""
+
+
+def add_default_command(parser):
+    """
+    Sets the default command to run when none is provided.
+    """
+    class default_command():
+        def run(args):
+            parser.print_help()
+            return 2
+
+    parser.set_defaults(__command__ = default_command)
+
